Update README.md
For my project, I have decided to create a research website which targets university students as users. The purpose of this website is to make the research process enjoyable for students, to increase their academic ability; providing a trustworthy platform they can rely on. To make the research experience for students more efficient, I designed a platform that is tailored to their needs. Therefore, it is different in comparison to the majority of research websites as its creative direction is user centred to emote amusement rather than seriousness
PLEASE SEE UPLOADED FILES FOR CODE AND REPORT 
Final grade - 1st class